Changed the test project over to .net 4.5. Bugfixes in the SQL generation. Removed "Rowid" from one of the default select columns. Fixed autoincrement in the SqlDatabaseGen class.
git-tfs-id: [https://djgosnell.visualstudio.com/DefaultCollection/]$/DtronixModel;C117

+++ DtxModel/DtxModeler/Generator/Sqlite/SqliteTypeTransformer.cs
@@ -7,7 +7,7 @@ namespace DtxModeler.Generator.Sqlite {
 	class SqliteTypeTransformer : TypeTransformer {
 		public SqliteTypeTransformer() {
 			types = new TypeTransformerType[]{
-				new TypeTransformerType("System.Int64", "BIGINT"),
+				new TypeTransformerType("System.Int64", "INTEGER"),
 				new TypeTransformerType("System.Int16", "SMALLINT"),
 				new TypeTransformerType("System.Int32", "INTEGER"),
 				/*new TypeTransformerType("System.UInt16", "SMALLINT"),
